@@ -15,7 +15,7 @@
  */
 package com.truizlop.sectionedrecyclerview;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.ViewGroup;
 
 /**
