@@ -15,7 +15,8 @@
  */
 package com.truizlop.sectionedrecyclerview;
 
-import android.support.v7.widget.GridLayoutManager;
+
+import androidx.recyclerview.widget.GridLayoutManager;
 
 /**
  * A SpanSizeLookup to draw section headers or footer spanning the whole width of the RecyclerView
