@@ -15,12 +15,13 @@
  */
 package com.truizlop.sectionedrecyclerview;
 
-import android.support.annotation.IdRes;
-import android.support.annotation.LayoutRes;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.annotation.IdRes;
+import androidx.annotation.LayoutRes;
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * An extension of SectionedRecyclerViewAdapter for simple sectioned RecyclerViews. In most cases,
